+++ src/Callbacks/CommaDelimiterSkuRelationsValidatorCallback.php
@@ -65,39 +65,19 @@ class CommaDelimiterSkuRelationsValidatorCallback extends ArrayValidatorCallback
             array_push($skuErrors, $value);
         }
         if (count($skuErrors) > 0) {
-            if (!$this->getSubject()->isStrictMode()) {
-                $message =  sprintf(
-                    'Found invalid SKUs "%s" to be related to %s product with SKU "%s"',
-                    implode(',', $skuErrors),
-                    $rowProductType,
-                    $rowSku
-                );
-                $this->getSubject()
-                    ->getSystemLogger()
-                    ->warning($this->getSubject()->appendExceptionSuffix($message));
-                $this->getSubject()->mergeStatus(
-                    array(
-                        RegistryKeys::NO_STRICT_VALIDATIONS => array(
-                            basename($this->getSubject()->getFilename()) => array(
-                                $this->getSubject()->getLineNumber() => array(
-                                    $attributeCode  => $message
-                                )
-                            )
-                        )
-                    )
-                );
+            $message =  sprintf(
+                'Found invalid SKUs "%s" to be related to %s product with SKU "%s"',
+                implode(',', $skuErrors),
+                $rowProductType,
+                $rowSku
+            );
+
+            if ($this->hasHandleStrictMode($attributeCode, $message)) {
                 return;
             }
 
             // throw an exception if the value is NOT in the array and strict mode on
-            throw new \InvalidArgumentException(
-                sprintf(
-                    'Found invalid SKUs "%s" to be related to %s product with SKU "%s"',
-                    implode(',', $skuErrors),
-                    $rowProductType,
-                    $rowSku
-                )
-            );
+            throw new \InvalidArgumentException($message);
         }
     }